Make this_session::set_id() public
In case the functions to dispatch are further distributed to some
downstream threads to not block receiver threads, the handler has to
forward the session id and therefore, needs to be able to access the
setters of this_session object.
In addition this_seesion().set_id() is called before the connection
callbacks.